requirements.txt
apache-airflow[celery,s3]==1.10.2
requests==2.21.0
